+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 1.x branch
 ## 1.5 branch
 ### 1.5.0
-* updated for `cakephp` 4.
+* updated for `cakephp` 4 and `phpunit` 8.
 
 ## 1.4 branch
 ### 1.4.3