+++ CHANGELOG.md
@@ -1,9 +1,7 @@
 # 1.x branch
 ## 1.4 branch
-### 1.4.3
-* removed useless `AssetsCreator::filename()` method.
-
 ### 1.4.2
+* removed useless `AssetsCreator::filename()` method;
 * added [API](//mirko-pagliai.github.io/cakephp-assets).
 
 ### 1.4.1